1Research_Analyst_CrewAI/pyproject.toml
[project]
name = "1research-analyst-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "crewai>=0.157.0",
    "crewai-tools>=0.60.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
